backend/requirements.txt
anyio==4.12.1
asgiref==3.11.0
Django==6.0.1
django-environ==0.12.0
django-filter==25.2
djangorestframework==3.16.1
idna==3.11
Markdown==3.10.1
psycopg2-binary==2.9.11
python-decouple==3.8
sqlparse==0.5.5
typing_extensions==4.15.0
watchfiles==1.1.1
qrcode[pil]==7.4.2
user-agents==2.2.0
django-cors-headers
